+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Upgrade the SDK to fully support Pydantic V2
+
 ## 1.22.3 - 2026-02-20
 
 ### Fixed